refactoring application structure
