+++ src/components/bootstrap/bootstrap.tsx
@@ -1,5 +1,9 @@
 import '../../../node_modules/bootstrap/dist/css/bootstrap.css';
 import * as React from 'react';
+import { connect } from 'react-redux';
+import { Credentials, CredentialsInterface } from '../../services/credentials';
+import { GithubApi } from '../../services/github-api';
+import { GITHUB_STORE_ACTIONS } from '../../stores/github-store';
 import Footer from '../footer/footer';
 import Header from '../header/header';
 import Navigation from '../navigation/navigation';
@@ -7,9 +11,27 @@ import UserDetails from '../user-details/user-details';
 
 // TODO any
 class Bootstrap extends React.Component<any, any> {
+  private credentials: CredentialsInterface;
+  private githubApi: any;
 
   constructor(props) {
     super(props);
+
+    this.credentials = new Credentials().getCredentials();
+    this.githubApi = new GithubApi(this.credentials);
+
+    this.getUserDetails();
+  }
+
+  private getUserDetails() {
+    let dispatch = this.props.dispatch;
+
+    this.githubApi.getUserDetails().then((response: any) => {
+      dispatch({
+        type: GITHUB_STORE_ACTIONS.GET_USER_DETAILS,
+        userDetails: response
+      })
+    });
   }
 
   render() {
@@ -48,4 +70,4 @@ class Bootstrap extends React.Component<any, any> {
   }
 }
 
-export default Bootstrap;
+export default connect()(Bootstrap);
