+++ src/components/user-details/user-details.tsx
@@ -1,9 +1,7 @@
 import './user-details.css';
 import * as React from 'react';
 import { connect } from 'react-redux';
-import { Credentials, CredentialsInterface } from '../../services/credentials';
-import { GithubApi, GithubIssuesInterface, GithubRepoInterface } from '../../services/github-api';
-import { GITHUB_STORE_ACTIONS } from '../../stores/github-store';
+
 
 function mapStateToProps(state) {
   return {
@@ -14,32 +12,13 @@ function mapStateToProps(state) {
 
 //TODO any
 class UserDetails extends React.Component<any, any> {
-  private credentials: CredentialsInterface;
-  private githubApi: any;
-
   constructor(props) {
     super(props);
 
-    this.credentials = new Credentials().getCredentials();
-    this.githubApi = new GithubApi(this.credentials);
     this.state = {
       avatar: '',
       username: ''
     };
-
-    this.getUserDetails();
-  }
-
-
-  private getUserDetails() {
-    let dispatch = this.props.dispatch;
-
-    this.githubApi.getUserDetails().then((response: any) => {
-      dispatch({
-        type: GITHUB_STORE_ACTIONS.GET_USER_DETAILS,
-        userDetails: response
-      })
-    });
   }
 
   componentWillReceiveProps(nextProps) {
